[touch: 34] Modified series part ordering (ordering including published posts count); adding pending posts to unpublished count; series parts reorder on posts status change.

+++ orgSeries-taxonomy.php
@@ -98,12 +98,26 @@ function set_series_order($postid = 0, $series_part = 0, $series_id, $is_publish
 	$post_ids_in_series = get_objects_in_term($series_id, 'series');
 	$series_posts = array(); 
  	$series_posts = get_series_order($post_ids_in_series, $postid, $series_id, true, false); 
- 	$total_posts = count($series_posts) + 1;
+ 	
+	if ($is_published) {
+		$total_posts = count($series_posts) + 1;
+	}
+	else{
+		$total_p_posts = 0;
+		foreach($series_posts as $s_p){
+			$s_pid = $s_p['id'];
+			$spost_status = get_post($s_pid)->post_status;
+			if($spost_status != 'draft' && $spost_status != 'future' && $spost_status!='pending')
+				$total_p_posts++;
+		}
+		$total_posts = $total_p_posts + 1;
+	}
 	
-	if (!isset($total_posts) || ($total_posts < $series_part) || $series_part ==  0 || $total_posts == 1) {
-		if ($total_posts >=1) $series_part = $total_posts;
+	if (!isset($total_posts) || (($total_posts < $series_part) && !$is_published) || $series_part ==  0 || $total_posts == 1) {
+		if ($total_posts >=1) 
+			$series_part = $total_posts;
 	} 
-	
+            
 	$ticker = 1;
 	$count = $total_posts;
 	if ($count >= 1) {
@@ -112,14 +126,23 @@ function set_series_order($postid = 0, $series_part = 0, $series_id, $is_publish
 			$spostid = $sposts['id'];
 			$spost_status = get_post($spostid)->post_status;
 			$is_was_rise = FALSE;
-			$is_was_drop = FALSE;
+                        
+			if ($spost_status != 'draft' && $spost_status != 'future' && $spost_status != 'pending')
+				$not_published = FALSE;
+			else
+				$not_published = TRUE;
 			
-			if($spost_status != 'draft' && $spost_status != 'future' && !$is_published)
+			if(!$not_published && !$is_published)
 				$spost_pchange = TRUE;
 			else 
 				$spost_pchange = FALSE;
 			
-			if (( $ticker >= 1) && ( $series_part > 2 ) &&  ( ($series_part - $currentpart)  >= 1) && $drop  && $spost_pchange )  {
+			if (($currentpart >= $count || $currentpart != $ticker) && ($currentpart > 1) && (($currentpart - $oldpart) > 1) && !$not_published){
+				$newpart = ($currentpart - 1);
+				$drop = TRUE;
+			}
+			
+			if (( $ticker >= 1) && ( $series_part > 2 ) && ( ($series_part - $currentpart)  >= 1) && $drop  && $spost_pchange )  {
 				$newpart = ($currentpart - 1);
 				$drop = TRUE;
 			}
@@ -136,8 +159,8 @@ function set_series_order($postid = 0, $series_part = 0, $series_id, $is_publish
 				
 			if ( ($series_part == $currentpart) && ( $series_part <= $count ) && ( $series_part > 1 ) && ($series_part != 2 ) && $drop  && $spost_pchange ) 
 				$newpart = ($currentpart - 1);
-				
-			if (( (($series_part == 1 ) && ($series_part >= $currentpart)) || (( $series_part == $currentpart )  && !$drop && ($currentpart - $oldpart) < 2) || (( $series_part < $currentpart ) && ( $currentpart == $oldpart ) && !$drop && ($currentpart != $count)) ) && $spost_pchange ) {
+			
+			if (( (($series_part == 1 ) && ($series_part >= $currentpart)) || (( $series_part == $currentpart ) && !$drop && ($currentpart - $oldpart) < 2) || (( $series_part < $currentpart ) && ( $currentpart == $oldpart ) && !$drop && ($currentpart != $count)) ) && $spost_pchange && $currentpart != $count) {
 				$newpart = ($currentpart + 1);
 				$rise = TRUE;
 				$is_was_rise = TRUE;
@@ -159,7 +182,7 @@ function set_series_order($postid = 0, $series_part = 0, $series_id, $is_publish
 			if (!isset($newpart)) 
 				$newpart = $currentpart;
 				
-			if ( isset($oldpart) && ($newpart - $oldpart) > 1 && !$drop && !$rise && ($newpart != ($count + 1) )  && $spost_pchange ) {
+			if ( isset($oldpart) && ($newpart - $oldpart) > 1 && !$drop && !$rise && ($newpart != ($count + 1)) && $spost_pchange ) {
 				$newpart = ($currentpart - 1);
 				$drop = TRUE;
 			}
@@ -170,7 +193,6 @@ function set_series_order($postid = 0, $series_part = 0, $series_id, $is_publish
 			$ticker++;
 			$oldpart = $newpart;
 			unset($newpart);
-			
 		}
 	}
 	$series_part_key = apply_filters('orgseries_part_key', SERIES_PART_KEY, $series_id);
@@ -184,10 +206,9 @@ function wp_reset_series_order_meta_cache ($post_id = 0, $series_id = 0, $reset 
 	if ( 0 == $series_id ) return false; //post is not a part of a series so no need to waste cycles.
 	
 	$post_ids_in_series = get_objects_in_term($series_id, 'series');
-	$spost_status = get_post($spostid)->post_status;
 	
 	$addvalue = 1;
-	
+        
 	$series_posts = get_series_order($post_ids_in_series, $post_id, $series_id, true, false);
 	$series_part_key = apply_filters('orgseries_part_key', SERIES_PART_KEY, $series_id);
 	
@@ -205,7 +226,8 @@ function wp_reset_series_order_meta_cache ($post_id = 0, $series_id = 0, $reset 
 	}
 	
 	foreach ($series_posts as $spost) {
-		if($spost_status != 'draft' && $spost_status != 'future'){
+                                    $spost_status = get_post($spost['id'])->post_status;
+		if($spost_status != 'draft' && $spost_status != 'future' && $spost_status != 'pending'){
 			$newpart = $addvalue;
 			delete_post_meta($spost['id'], $series_part_key);
 			add_post_meta($spost['id'], $series_part_key, $newpart);
@@ -443,11 +465,14 @@ function wp_set_post_series( $post_ID = 0, $post, $series_id = array(), $dont_sk
 			
 	/*print_r($series_part);
 	exit;/**/
-		//if we don't have any changes in the series or series part info then let's get out and save time.
+		//if we don't have any changes in the series or series part info (or series post status) then let's get out and save time.
+		$p_status = $post->post_status;
+		if($p_status != 'draft' && $p_status != 'future' && $p_status != 'pending')
+			$ispublished = TRUE;
 		$count = count($post_series);
 		$c_chk = 0;
 		foreach ( $post_series as $ser ) {
-			if (in_array($ser, $old_series) && $series_part[$ser] == wp_series_part($post_ID, $ser) && !$dont_skip ) {
+			if (in_array($ser, $old_series) && $series_part[$ser] == wp_series_part($post_ID, $ser) && !$dont_skip && $ispublished) {
 				$c_chk++;
 				continue;
 			} else {
@@ -492,18 +517,27 @@ function wp_set_post_series( $post_ID = 0, $post, $series_id = array(), $dont_sk
 	
 	if ( empty($p_ser_edit) ) return; //let's get out we've done everything we need to do.
 	if ( $success ) {
+		if($p_status == 'draft' || $p_status == 'future' || $p_status == 'pending')
+			$is_published = TRUE;
 		foreach ( $p_ser_edit as $ser_id ) {
 			if ( empty($series_part[$ser_id]) ) {
 				$s_pt = wp_series_part($post_ID, $ser_id);
 				if ( !$series_part ) $series_part = 0;
-			} else {
+			} 
+			
+			//If post is not published its part stays as set by user
+			elseif(!$is_published) {
 				$s_pt = $series_part[$ser_id];
+			}
+			else{
+				if (isset($_GET['submit']))
+					$set_spart = $_GET['series_part'];
+				else
+					$set_spart =  $_POST['series_part'];
+				$s_pt = (int) implode($set_spart);
 			}
 			/*print_r($s_pt);
 			exit;/**/
-			$p_status = $post->post_status;
-			if($p_status == 'draft' || $p_status == 'future')
-				$is_published = true;
 			set_series_order($post_ID, $s_pt, $ser_id, $is_published);
 		}
